【修正】ワーニングを消すため、変数の型統一
・理由：ワーニングが出ていた為
・対応内容：int→floatに変更
・影響範囲：問題なく起動した
・備考：

+++ ShootingGame_0519/HitPointCompornent.cpp
@@ -2,8 +2,7 @@
 #include "GameObject.h"
 #include <algorithm>
 
-HitPointComponent::HitPointComponent(int maxHP)
-	: m_maxHp(maxHP), m_hp(maxHP)
+HitPointComponent::HitPointComponent(float maxHP): m_maxHp(maxHP), m_hp(maxHP)
 {
 	//コンストラクタ
 }
